+++ packages/sdk/src/types/transactionEvent.ts
@@ -9,7 +9,7 @@ import type {
     OpenStatusText,
     ReleaseSchedule,
 } from '../types.js';
-import { TokenId } from '../plt/types.js';
+import * as PLT from '../plt/types.js';
 
 import type * as AccountAddress from './AccountAddress.js';
 import type * as CcdAmount from './CcdAmount.js';
@@ -61,8 +61,8 @@ export enum TransactionEventTag {
     Resumed = 'Resumed',
     Updated = 'Updated',
     Upgraded = 'Upgraded',
-    TokenGovernanceUpdate = 'TokenGovernanceUpdate',
-    TokenHolderUpdate = 'TokenHolderUpdate',
+    TokenGovernance = 'TokenGovernance',
+    TokenHolder = 'TokenHolder',
 }
 
 export type TransactionEvent =
@@ -94,7 +94,9 @@ export type TransactionEvent =
     | BakerSetFinalizationRewardCommissionEvent
     | BakerSetBakingRewardCommissionEvent
     | BakerSetTransactionFeeCommissionEvent
-    | UpdateEnqueuedEvent;
+    | UpdateEnqueuedEvent
+    | TokenHolderEvent
+    | TokenGovernanceEvent;
 
 // Contract Events
 
@@ -362,26 +364,28 @@ export interface UpdateEnqueuedEvent {
 * Token (PLT) events originating from governance transactions.
 */
 export type TokenGovernanceEvent = {
-    tag: TransactionEventTag.TokenGovernanceUpdate;
-    /** The ID of the token. */
-    id: TokenId.Type;
-    /** The type of the event. */
-    type: string;
-    /** The CBOR encoded details of the event. These should be decoded using the appropriate CDDL schema. */
-    details: HexString; // TODE: maybe create a newtype here to enforce invariants?
+    tag: TransactionEventTag.TokenGovernance;
+    /** The token ID of the token the event originates from */
+    tokenId: PLT.TokenId.Type,
+    /** The specific type of update made to the token instance */
+    updateType: string,
+    /** The CBOR encoded details of the update. The details might vary depending on the PLT module reference used to
+     * instantiate the token. */
+    details: PLT.TokenEvent.Type,
 };
 
 /**
 * Token (PLT) events originating from account transactions.
 */
 export type TokenHolderEvent = {
-    tag: TransactionEventTag.TokenHolderUpdate;
-    /** The ID of the token. */
-    id: TokenId.Type;
-    /** The type of the event. */
-    type: string;
-    /** The CBOR encoded details of the event. These should be decoded using the appropriate CDDL schema. */
-    details: HexString; // TODE: maybe create a newtype here to enforce invariants?
+    tag: TransactionEventTag.TokenHolder;
+    /** The token ID of the token the event originates from */
+    tokenId: PLT.TokenId.Type,
+    /** The specific type of update made to the token instance */
+    updateType: string,
+    /** The CBOR encoded details of the update. The details might vary depending on the PLT module reference used to
+     * instantiate the token. */
+    details: PLT.TokenEvent.Type,
 };
 
 export type ContractTraceEvent = ResumedEvent | InterruptedEvent | UpdatedEvent | UpgradedEvent | TransferredEvent;